Student  Tests

+++ backend/__tests__/util/truncate.js
@@ -3,10 +3,11 @@ import database from '../../src/database';
 export default function truncate() {
   return Promise.all(
     Object.keys(database.connection.models).map(key => {
-      return database.connection.models[key].destroy({
-        truncate: true,
-        force: true,
-      });
+      if (key !== 'User')
+        return database.connection.models[key].destroy({
+          truncate: true,
+          force: true,
+        });
     })
   );
 }
